backend/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
SQLAlchemy = "^2.0.13"
tenacity = "^8.2.2"
asyncpg = "^0.27.0"
fastapi-users = "^11.0.0"
fastapi-users-db-sqlalchemy = "^5.0.0"
httpx-oauth = "^0.11.2"
elevenlabs = "^0.2.16"
alembic = "^1.11.1"
redis = "^4.5.5"
randomname = "^0.2.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
